[LOGMGR-218] Follows up on the previous commit which could potentially initialize JBoss Modules too early for some use cases. One specific use case is WildFly embedded.
